modificações no service, dto e exception do Feedback, teste do feedbackService feito

+++ src/main/java/com/Weimar_Jr/SistemaDeSugestoesEFeedbackDeLivrosDaBibliotecaEscolar/EntidadesDTO/Biblioteca/Feedback/AtualizarFeedbackDTORequest.java
@@ -6,7 +6,7 @@ public record AtualizarFeedbackDTORequest(
         @JsonProperty("comentario")
         String comentario,
         @JsonProperty("avaliacao")
-        Integer avaliacao,
+        Double avaliacao,
         @JsonProperty("nome_do_aluno_visivel")
         Boolean nomeDoAlunoVisivel
 )
